востановил InfoController, чтобы не нагромождать SearchController
