+++ Web/StoreProduct.Web/Controllers/SearchController.cs
@@ -67,34 +67,5 @@ namespace StoreProduct.Web.Controllers
 
             return View("Index", model);
         }
-
-
-
-        // получание продукта из репозитория и вывод его полей, запрос по Id
-        public IActionResult InfoProduct(int id)
-        {
-            var product = productService.GetById(id);
-
-            var model = new ProductModel
-            {
-                ProductId = product.Id,
-                MakerId = product.IdMaker,
-                ProductTitle = product.Title,
-                MakerTitle = makerRepository.GetById(product.IdMaker).Title,
-                Category = product.Category,
-                Price = product.Price,
-                Description = product.Description
-            };
-
-            return View("InfoProduct", model);
-        }
-
-        // получание производителя из репозитория и вывод его полей, запрос по Id
-        public IActionResult InfoManufacture(int id)
-        {
-            var manufacture = makerRepository.GetById(id);
-
-            return View("InfoManufacture", manufacture);
-        }
     }
 }
