+++ infastrucrture/Store.Memory/ProductRepository.cs
@@ -16,7 +16,7 @@ namespace Store.Memory
                 //db.Products.Add(new Product("говядина", manufactures.GetById(3), "мясо", 30m, "мясо из говядины и телятины"));
                 //db.Products.Add(new Product("свинина", manufactures.GetById(4), "мясо", 40m, "мясо из свинины"));
 
-                db.SaveChanges();
+                //db.SaveChanges();
 
                 products = db.Products.ToList();
             }
